The npm script names must be different or vscode only shows one!
This is by design. An extension provided task, such as a task of "type": "npm" is identified by it task-type-specific properties (for npm tasks, this is script). Tasks are not identified by label, a label is just a nice human readable string.

Since this is very easily worked around by having a separate npm script in your package.json I wouldn't make any changes in the behavior. However, it is not nice that the first time you learn that you can't have two of the same task is when you try to run a task that depends on it. I will add an error earlier, so that it's clear what's happening when the task list shows.